# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MikroTik Centralized Fleet Management API
  version: 1.20.0
...
servers:
- url: /api/v1
paths:
...
  /devices:
    get:
      summary: List devices
      tags:
      - Inventory
      parameters:
      - *id001
      - in: query
        name: ship_id
        schema:
          type: string
          format: uuid
      - in: query
        name: role
        schema:
          type: string
          enum: &id033
          - EDGE
          - CORE
          - SWITCH
          - AP
          - CPE
      - in: query
        name: status
        schema:
          type: string
          enum: &id034
          - UNKNOWN
          - ONLINE
          - DEGRADED
          - OFFLINE
          - MAINTENANCE
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id004
    post:
      summary: Create device
      tags:
      - Inventory
      parameters:
      - *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Device'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
              example: &id008
                data:
                  id: ce067c3e-16c4-4211-969c-0a10e8a5e407
                  ship_id: c328fbf6-74f0-445d-9147-e60477a2ab2c
                  code: CCR-CORE-1
                  name: CCR-CORE-1
                  role: CORE
                  model: CCR2004
                  serial: null
                  architecture: null
                  routeros_version: null
                  device_mode: null
                  api_transport: REST
                  credential_ref: null
                  mgmt_endpoint_ref: null
                  status: UNKNOWN
                  last_seen_at: null
                  poll_interval_s: 60
                  created_at: '2026-08-25T01:15:45.925Z'
                  updated_at: '2026-08-25T01:15:45.925Z'
                meta:
                  request_id: req_01M0V7SNC0R6HVERVZWW8FD8KP
                  generated_at: '2026-08-25T01:15:45.929Z'
                error: null
        '400': *id003
        '403': *id004
        '404':
          description: SHIP_NOT_FOUND
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
  /devices/{deviceId}:
    get:
      summary: Get device detail
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Device'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
              example: *id008
        '403': *id004
        '404': *id006
    patch:
      summary: Update device
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Device'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id003
        '403': *id004
        '404': *id006
    delete:
      summary: Delete device (soft delete)
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '204':
          description: Deleted
        '403': *id004
        '404': *id006
  /devices/{deviceId}/traffic:
    get:
      summary: Traffic theo bucket cho mot thiet bi (WAN total + interface rate theo thoi gian)
      description: Tra 200 kem data_status INSUFFICIENT_DATA khi chua co interface_counter_deltas nao cho thiet bi nay, khong throw loi -- giong ho Global/Area/Ship Dashboard, khac Reconciliation.
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DeviceTraffic'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id004
        '404': *id006
  /devices/{deviceId}/interfaces/traffic:
    get:
      summary: Bandwidth + tong data THEO TUNG INTERFACE cho mot thiet bi
      description: Khac GET .../traffic (chi gop theo 4 nhom WAN/CREW/BUSINESS/MANAGEMENT) -- endpoint nay tra ve tung interface rieng de ve 1 bieu do bandwidth nhieu duong + tong data/interface. month_total reset tu nhien vao 00:00 UTC ngay 1 hang thang (tinh tu dau thang UTC den hien tai, khong phai gia tri cache bi cron xoa).
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      - $ref: '#/components/parameters/DashboardFrom'
      - $ref: '#/components/parameters/DashboardTo'
      - $ref: '#/components/parameters/DashboardTimezone'
      - $ref: '#/components/parameters/DashboardGranularity'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DeviceInterfaceTraffic'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id004
        '404': *id006
  /devices/{deviceId}/push-key:
    post:
      summary: Cap push API key moi cho thiet bi (model push 1 chieu)
      description: Sinh key that ngau nhien, luu sha256(key) vao devices.api_key_hash, tra key that mot lan duy nhat trong response. Cap lai se vo hieu key cu ngay lap tuc.
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DevicePushKey'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id004
        '404': *id006
    delete:
      summary: Thu hoi push API key cua thiet bi
      description: Xoa api_key_hash/api_key_issued_at -- lan push ke tiep bang key cu se nhan DEVICE_PUSH_NOT_CONFIGURED (409).
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '204':
          description: Revoked
        '403': *id004
        '404': *id006
  /devices/{deviceId}/radius-secret:
    post:
      summary: Cap RADIUS secret moi cho thiet bi
      description: Sinh secret that ngau nhien qua EnvSecretStore (TAM THOI, ADR-05 -- xem backend/src/libs/secrets), tu dat credential_ref tro toi bien vua tao, thu hoi bien cu neu co. Secret that tra ve DUY NHAT 1 LAN trong response nay; RADIUS Accounting-Request ke tiep tu router dung duoc ngay khong can restart server.
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DeviceRadiusSecret'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '403': *id004
        '404': *id006
    delete:
      summary: Thu hoi RADIUS secret cua thiet bi
      description: Xoa credential_ref/radius_secret_issued_at va bien env tuong ung -- goi Accounting-Request ke tiep tu router dung secret cu se bi RADIUS server tu choi (khong xac thuc duoc Request Authenticator).
      tags:
      - Inventory
      parameters:
      - *id001
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      responses:
        '204':
          description: Revoked
        '403': *id004
        '404': *id006
  /devices/{deviceId}/telemetry-push:
    post:
      summary: MikroTik tu day telemetry len server (khong qua actor guard)
      description: Router tu goi vao moi 5 phut qua /system scheduler + /tool fetch, xac thuc bang header X-Device-Api-Key (KHONG PHAI x-actor-permissions/AuthStubGuard). Tai dung nguyen pipeline POST /telemetry/ingest de dedup/normalize.
      tags:
      - Inventory
      parameters:
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
        description: ''
      - in: header
        name: X-Device-Api-Key
        required: true
        description: Key that do POST .../push-key cap -- server so sanh sha256(key) voi devices.api_key_hash.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTelemetryPushRequest'
      responses:
        '201':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DeviceTelemetryPushResult'
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    allOf:
                    - $ref: '#/components/schemas/ErrorDetails'
                    nullable: true
        '400': *id003
        '401':
          description: DEVICE_PUSH_UNAUTHORIZED -- thieu hoac sai X-Device-Api-Key
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
        '404': *id006
        '409':
          description: DEVICE_PUSH_NOT_CONFIGURED -- thiet bi chua duoc cap push API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    nullable: true
                  meta:
                    $ref: '#/components/schemas/BaseMeta'
                  error:
                    $ref: '#/components/schemas/ErrorDetails'
...
components:
  parameters:
...
    DashboardFrom:
      in: query
      name: from
      required: false
      schema:
        type: string
        format: date-time
      description: RFC3339 period start. Defaults to 24 hours before `to`.
    DashboardTo:
      in: query
      name: to
      required: false
      schema:
        type: string
        format: date-time
      description: RFC3339 period end. Defaults to now.
    DashboardTimezone:
      in: query
      name: timezone
      required: false
      schema:
        type: string
      description: IANA timezone used for period bucketing.
    DashboardGranularity:
      in: query
      name: granularity
      required: false
      schema:
        type: string
        enum:
        - 1m
        - 5m
        - 1h
        - 1d
        default: 1h
...
  schemas:
    BaseMeta:
      type: object
      properties:
        request_id:
          type: string
        generated_at:
          type: string
          format: date-time
        data_from:
          type: string
          format: date-time
        data_to:
          type: string
          format: date-time
        data_freshness_seconds:
          type: integer
          nullable: true
        freshness_by_source:
          type: object
          additionalProperties:
            type: integer
            nullable: true
        page:
          type: object
          nullable: true
          description: Chi xuat hien tren cac endpoint list co phan trang (cursor hoac offset).
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
            cursor_next:
              type: string
              nullable: true
            has_more:
              type: boolean
        warnings:
          type: array
          nullable: true
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              details:
                type: object
                additionalProperties: true
...
    DeviceTraffic:
      type: object
      description: WAN total va rate theo bucket cho mot thiet bi, tu interface_counter_deltas (WAN_INPUT). data_status INSUFFICIENT_DATA khi chua co delta nao -- khong phai loi, chi la chua co telemetry.
      properties:
        device_id:
          type: string
          format: uuid
        period:
          $ref: '#/components/schemas/DashboardPeriod'
        units:
          $ref: '#/components/schemas/DashboardUnits'
        data_status:
          type: string
          enum:
          - AVAILABLE
          - INSUFFICIENT_DATA
        wan:
          type: object
          description: Tong download/upload WAN trong ca khoang period (khong phai 1 bucket).
          properties:
            download_bytes:
              type: integer
              nullable: true
            upload_bytes:
              type: integer
              nullable: true
        reconciliation:
          type: object
          description: Doi soat tong data thuc te da len server qua WAN so voi tong da gan nhan CREW/BUSINESS/MANAGEMENT trong cung khoang thoi gian -- cung cong thuc computeGap() dung o Reconciliation cap tau, loc theo device_id. gap duong nghia la co traffic WAN chua duoc gan vao interface nao (con NONE), khong phai loi he thong.
          properties:
            counted:
              type: object
              description: Tong CREW_ACCESS + BUSINESS_ACCESS + MANAGEMENT da dem duoc -- null neu chua co interface nao trong 3 nhom nay co delta.
              properties:
                download_bytes:
                  type: integer
                  nullable: true
                upload_bytes:
                  type: integer
                  nullable: true
            gap:
              type: object
              description: gap = wan - counted. Duong = con traffic WAN chua gan nhom; am (hiem) = counted vuot wan (vd loi dem 2 lan).
              properties:
                download_bytes:
                  type: integer
                  nullable: true
                download_pct:
                  type: number
                  nullable: true
                upload_bytes:
                  type: integer
                  nullable: true
                upload_pct:
                  type: number
                  nullable: true
        points:
          type: array
          items:
            type: object
            properties:
              bucket:
                type: string
                format: date-time
              wan:
                type: object
                nullable: true
                properties:
                  download_bytes:
                    type: integer
                  upload_bytes:
                    type: integer
              crew:
                type: object
                nullable: true
                properties:
                  download_bytes:
                    type: integer
                  upload_bytes:
                    type: integer
              business:
                type: object
                nullable: true
                properties:
                  download_bytes:
                    type: integer
                  upload_bytes:
                    type: integer
              management:
                type: object
                nullable: true
                properties:
                  download_bytes:
                    type: integer
                  upload_bytes:
                    type: integer
    DeviceInterfaceTraffic:
      type: object
      description: Bandwidth theo bucket + tong data THEO TUNG INTERFACE (khac DeviceTraffic chi gop theo 4 nhom WAN/CREW/BUSINESS/MANAGEMENT). month_total reset tu nhien vao 00:00 UTC ngay 1 hang thang -- tinh tu dau thang UTC hien tai den luc goi API, khong phai gia tri luu san bi reset boi cron.
      properties:
        device_id:
          type: string
          format: uuid
        period:
          $ref: '#/components/schemas/DashboardPeriod'
        units:
          $ref: '#/components/schemas/DashboardUnits'
        data_status:
          type: string
          enum:
          - AVAILABLE
          - INSUFFICIENT_DATA
        interfaces:
          type: array
          items:
            type: object
            properties:
              interface_id:
                type: string
                format: uuid
              name:
                type: string
              accounting_group:
                type: string
                enum:
                - WAN_INPUT
                - CREW_ACCESS
                - BUSINESS_ACCESS
                - MANAGEMENT
                - NONE
              points:
                type: array
                items:
                  type: object
                  properties:
                    bucket:
                      type: string
                      format: date-time
                    rx_bytes:
                      type: integer
                    tx_bytes:
                      type: integer
              month_total:
                type: object
                description: Tong cong don tu 00:00 UTC ngay 1 cua thang hien tai den luc goi API -- null neu interface nay chua co delta nao trong thang.
                properties:
                  from:
                    type: string
                    format: date-time
                  to:
                    type: string
                    format: date-time
                  rx_bytes:
                    type: integer
                    nullable: true
                  tx_bytes:
                    type: integer
                    nullable: true
                  total_bytes:
                    type: integer
                    nullable: true
              latest_sample:
                type: object
                nullable: true
                description: Nguyen van payload MikroTik gui lan gan nhat (counter CONG DON tu luc router boot, khong phai delta) -- packets/errors/drops chi khac null khi router thuc su gui truong do.
                properties:
                  observed_at:
                    type: string
                    format: date-time
                  rx_bytes:
                    type: integer
                  tx_bytes:
                    type: integer
                  rx_packets:
                    type: integer
                    nullable: true
                  tx_packets:
                    type: integer
                    nullable: true
                  rx_errors:
                    type: integer
                    nullable: true
                  tx_errors:
                    type: integer
                    nullable: true
                  rx_drops:
                    type: integer
                    nullable: true
                  tx_drops:
                    type: integer
                    nullable: true
        last_push:
          type: object
          nullable: true
          description: Thong tin router tu bao o lan push gan nhat -- null neu chua tung nhan push nao.
          properties:
            identity:
              type: string
              nullable: true
              description: source_device_ref cua raw_telemetry_events gan nhat -- vd /system identity name MikroTik tu goi len.
            received_at:
              type: string
              format: date-time
...
    ErrorDetails:
      type: object
      description: Body loi chuan hoa -- AGENT_COLLABORATION.md Sec.5. Khong bao gio chua stack trace hay loi driver that.
      properties:
        code:
          type: string
          enum:
          - VALIDATION_FAILED
          - INVALID_TIME_RANGE
          - INSUFFICIENT_DATA
          - RECONCILIATION_UNAVAILABLE
          - UNAUTHENTICATED
          - FORBIDDEN
          - SCOPE_FORBIDDEN
          - AREA_NOT_FOUND
          - SHIP_NOT_FOUND
          - DEVICE_NOT_FOUND
          - INTERFACE_NOT_FOUND
          - ZONE_NOT_FOUND
          - ENDPOINT_NOT_FOUND
          - ALERT_NOT_FOUND
          - SERVICE_NOT_FOUND
          - RESOURCE_CONFLICT
          - AREA_HAS_SHIPS
          - SHIP_HAS_DEVICES
          - INTERFACE_DOUBLE_COUNT
          - LAST_HEALTHY_ENDPOINT
          - ENDPOINT_ALREADY_EXISTS
          - DATABASE_UNAVAILABLE
          - RATE_LIMITED
          - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
        retryable:
          type: boolean
        retry_after_seconds:
          type: integer
          nullable: true
      required:
      - code
      - message
      - retryable
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ship_id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        role:
          type: string
          enum: *id033
        model:
          type: string
          nullable: true
        serial:
          type: string
          nullable: true
        architecture:
          type: string
          nullable: true
        routeros_version:
          type: string
          nullable: true
        device_mode:
          type: string
          nullable: true
        api_transport:
          type: string
          enum: &id035
          - REST
          - API_SSL
          - API
        credential_ref:
          type: string
          nullable: true
          description: Tham chieu toi secret store (ADR-05), KHONG PHAI gia tri credential that.
        mgmt_endpoint_ref:
          type: string
          nullable: true
        status:
          type: string
          enum: *id034
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        poll_interval_s:
          type: integer
        ip_address: &deviceIpAddress
          type: string
          nullable: true
          description: Inventory that (khong phai secret) -- IP quan tri cua router, dung de dung script push mau. Khac credential_ref/mgmt_endpoint_ref.
        push_key_configured:
          type: boolean
          description: true neu thiet bi da duoc cap push API key (POST .../push-key). KHONG BAO GIO tra ve chinh key/hash -- giong secret_configured (ADR-05).
        push_key_issued_at:
          type: string
          format: date-time
          nullable: true
        radius_secret_configured:
          type: boolean
          description: true neu thiet bi da co credential_ref (POST .../radius-secret hoac PATCH tay). KHONG BAO GIO tra ve chinh secret that -- giong secret_configured (ADR-05).
        radius_secret_issued_at:
          type: string
          format: date-time
          nullable: true
        router_config_text: &deviceRouterConfigText
          type: string
          nullable: true
          description: Nguyen van RouterOS config (vd /export hide-sensitive) de tham khao/doi chieu -- server KHONG BAO GIO tu thuc thi hay day xuong router.
        router_config_updated_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DeviceCreateRequest:
      type: object
      required:
      - ship_id
      - code
      - name
      - role
      properties:
        ship_id:
          type: string
          format: uuid
        code: &id036
          type: string
          maxLength: 64
        name: &id037
          type: string
          maxLength: 255
        role: &id038
          type: string
          enum: *id033
        model: &id039
          type: string
          nullable: true
        serial: &id040
          type: string
          nullable: true
        architecture: &id041
          type: string
          nullable: true
        routeros_version: &id042
          type: string
          nullable: true
        api_transport: &id043
          type: string
          enum: *id035
          default: REST
        credential_ref: &id044
          type: string
          nullable: true
        mgmt_endpoint_ref: &id045
          type: string
          nullable: true
        poll_interval_s: &id046
          type: integer
          default: 60
        ip_address: *deviceIpAddress
        router_config_text: *deviceRouterConfigText
    DeviceUpdateRequest:
      type: object
      properties:
        code: *id036
        name: *id037
        role: *id038
        model: *id039
        serial: *id040
        architecture: *id041
        routeros_version: *id042
        api_transport: *id043
        credential_ref: *id044
        mgmt_endpoint_ref: *id045
        poll_interval_s: *id046
        ip_address: *deviceIpAddress
        router_config_text: *deviceRouterConfigText
        status:
          type: string
          enum: *id034
    DevicePushKey:
      type: object
      description: Ket qua cap/thu hoi push API key. api_key la key THAT, chi tra ve DUY NHAT 1 LAN trong response cua POST -- server chi luu sha256(api_key), khong doc lai duoc sau do.
      properties:
        api_key:
          type: string
          description: Key that (hex, 64 ky tu) -- dan vao script RouterOS ngay, se KHONG hien lai.
        issued_at:
          type: string
          format: date-time
    DeviceRadiusSecret:
      type: object
      description: Ket qua cap/thu hoi RADIUS secret. secret la gia tri THAT, chi tra ve DUY NHAT 1 LAN trong response cua POST -- server khong luu lai/log lai sau do (di qua EnvSecretStore, xem ADR-05).
      properties:
        credential_ref:
          type: string
          description: Con tro da tu dong gan vao devices.credential_ref, dang "env:TEN_BIEN" -- day la gia tri AN TOAN de hien thi lai (khong phai secret that).
        secret:
          type: string
          description: Secret that (hex, 48 ky tu) -- dan vao "/radius add ... secret=..." tren router ngay, se KHONG hien lai.
        issued_at:
          type: string
          format: date-time
    DeviceTelemetryPushRequest:
      type: object
      description: Body ma chinh MikroTik tu gui len (POST /devices/{deviceId}/telemetry-push, header X-Device-Api-Key thay cho x-actor-permissions) -- model push 1 chieu, scheduler RouterOS chay moi 5 phut.
      required:
      - interfaces
      properties:
        identity:
          type: string
          nullable: true
          description: Nhan dien thiet bi phia nguon (vd /system identity print), tuy chon.
        ip_address:
          type: string
          nullable: true
          description: IP nguon cua router luc gui -- neu co, server tu cap nhat devices.ip_address.
        interfaces:
          type: array
          minItems: 1
          items:
            type: object
            required:
            - name
            - rx_byte
            - tx_byte
            properties:
              name:
                type: string
                description: Ten interface dung nhu tren thiet bi (vd ether1) -- normalize resolve theo (device_id, name).
              rx_byte:
                type: number
                minimum: 0
              tx_byte:
                type: number
                minimum: 0
              rx_packet:
                type: number
                minimum: 0
                nullable: true
              tx_packet:
                type: number
                minimum: 0
                nullable: true
              rx_error:
                type: number
                minimum: 0
                nullable: true
              tx_error:
                type: number
                minimum: 0
                nullable: true
    DeviceTelemetryPushResult:
      type: object
      properties:
        accepted:
          type: integer
        duplicate:
          type: integer
        last_seen_at:
          type: string
          format: date-time
...